gh-1659 - Undo package renaming.
