+++ store-implementation/map-store/src/main/java/uk/gov/gchq/gaffer/mapstore/MapStore.java
@@ -22,12 +22,12 @@ import org.slf4j.LoggerFactory;
 import uk.gov.gchq.gaffer.commonutil.iterable.CloseableIterable;
 import uk.gov.gchq.gaffer.data.element.Element;
 import uk.gov.gchq.gaffer.data.element.id.EntityId;
-import uk.gov.gchq.gaffer.mapstore.operation.impl.AddElementsHandler;
-import uk.gov.gchq.gaffer.mapstore.operation.impl.CountAllElementsDefaultViewHandler;
-import uk.gov.gchq.gaffer.mapstore.operation.impl.GetAdjacentIdsHandler;
-import uk.gov.gchq.gaffer.mapstore.operation.impl.GetAllElementsHandler;
-import uk.gov.gchq.gaffer.mapstore.operation.impl.GetElementsHandler;
-import uk.gov.gchq.gaffer.mapstore.operation.impl.MapImpl;
+import uk.gov.gchq.gaffer.mapstore.impl.AddElementsHandler;
+import uk.gov.gchq.gaffer.mapstore.impl.CountAllElementsDefaultViewHandler;
+import uk.gov.gchq.gaffer.mapstore.impl.GetAdjacentIdsHandler;
+import uk.gov.gchq.gaffer.mapstore.impl.GetAllElementsHandler;
+import uk.gov.gchq.gaffer.mapstore.impl.GetElementsHandler;
+import uk.gov.gchq.gaffer.mapstore.impl.MapImpl;
 import uk.gov.gchq.gaffer.mapstore.operation.CountAllElementsDefaultView;
 import uk.gov.gchq.gaffer.operation.impl.add.AddElements;
 import uk.gov.gchq.gaffer.operation.impl.get.GetAdjacentIds;
