+++ store-implementation/map-store/src/main/java/uk/gov/gchq/gaffer/mapstore/impl/package-info.java
@@ -17,4 +17,4 @@
 /**
  * Implementation details for the MapImpl map-based data store.
  */
-package uk.gov.gchq.gaffer.mapstore.operation.impl;
+package uk.gov.gchq.gaffer.mapstore.impl;
