+++ store-implementation/map-store/src/test/java/uk/gov/gchq/gaffer/mapstore/utils/ElementClonerTest.java
@@ -21,7 +21,7 @@ import uk.gov.gchq.gaffer.commonutil.pair.Pair;
 import uk.gov.gchq.gaffer.commonutil.stream.Streams;
 import uk.gov.gchq.gaffer.mapstore.MapStore;
 import uk.gov.gchq.gaffer.mapstore.MapStoreProperties;
-import uk.gov.gchq.gaffer.mapstore.operation.impl.GetAllElementsHandlerTest;
+import uk.gov.gchq.gaffer.mapstore.impl.GetAllElementsHandlerTest;
 import uk.gov.gchq.gaffer.store.StoreException;
 
 import static org.junit.Assert.assertEquals;
